Add CustomContent maps to UserData and link them

+++ src/main/java/org/tndata/android/compass/model/CustomAction.java
@@ -36,7 +36,7 @@ public class CustomAction extends Action{
     private CustomGoal mGoal;
 
 
-    public int getCustomGoalId(){
+    public long getCustomGoalId(){
         return mCustomGoalId;
     }
